+++ auth/provider_session.go
@@ -109,10 +109,6 @@ func NewProviderSessionUpdate() *ProviderSessionUpdate {
 	return &ProviderSessionUpdate{}
 }
 
-func (p *ProviderSessionUpdate) HasUpdates() bool {
-	return p.OAuthToken != nil
-}
-
 func (p *ProviderSessionUpdate) Parse(parser structure.ObjectParser) {
 	if oauthTokenParser := parser.WithReferenceObjectParser("oauthToken"); oauthTokenParser.Exists() {
 		p.OAuthToken = oauth.NewToken()
@@ -125,6 +121,10 @@ func (p *ProviderSessionUpdate) Validate(validator structure.Validator) {
 	if p.OAuthToken != nil {
 		p.OAuthToken.Validate(validator.WithReference("oauthToken"))
 	}
+}
+
+func (p *ProviderSessionUpdate) IsEmpty() bool {
+	return p.OAuthToken == nil
 }
 
 func NewProviderSessionID() string {